rm old links
