+++ packages/client/src/links/httpBatchLink.ts
@@ -1,19 +1,15 @@
 import { AnyRouter, ProcedureType } from '@trpc/server';
-import { TRPCResponse } from '@trpc/server/rpc';
-import { TRPCClientError } from '../TRPCClientError';
 import { dataLoader } from '../internals/dataLoader';
-import { transformRPCResponse } from '../internals/transformRPCResponse';
-import { httpRequest } from '../internals/httpRequest';
-import { HTTPLinkOptions, TRPCLink } from './core';
-import { TRPCAbortError } from '../internals/TRPCAbortError';
+import { transformOperationResult } from './transformerLink';
+import { TRPCLink } from './core2';
+import { observable } from '../rx/observable';
+import { HTTPLinkOptions, httpRequest, ResponseShape } from './httpUtils';
 
 export function httpBatchLink<TRouter extends AnyRouter>(
   opts: HTTPLinkOptions,
 ): TRPCLink<TRouter> {
   const { url } = opts;
-  // initialized config
   return (runtime) => {
-    // initialized in app
     type Key = { id: number; path: string; input: unknown };
     const fetcher = (type: ProcedureType) => (keyInputPairs: Key[]) => {
       const path = keyInputPairs.map((op) => op.path).join(',');
@@ -28,42 +24,52 @@ export function httpBatchLink<TRouter extends AnyRouter>(
       });
 
       return {
-        promise: promise.then((res: unknown[] | unknown) => {
-          if (!Array.isArray(res)) {
-            return keyInputPairs.map(() => res);
-          }
-          return res;
+        promise: promise.then((res) => {
+          const resJSON = Array.isArray(res.json)
+            ? res.json
+            : keyInputPairs.map(() => res.json);
+
+          const result = resJSON.map((item) => ({
+            meta: res.meta,
+            json: item,
+          }));
+
+          return result;
         }),
         cancel,
       };
     };
-    const query = dataLoader<Key, TRPCResponse>(fetcher('query'));
-    const mutation = dataLoader<Key, TRPCResponse>(fetcher('mutation'));
-    const subscription = dataLoader<Key, TRPCResponse>(fetcher('subscription'));
+    const query = dataLoader<Key, ResponseShape>(fetcher('query'));
+    const mutation = dataLoader<Key, ResponseShape>(fetcher('mutation'));
+    const subscription = dataLoader<Key, ResponseShape>(
+      fetcher('subscription'),
+    );
 
     const loaders = { query, subscription, mutation };
-    return ({ op, prev, onDestroy }) => {
-      const loader = loaders[op.type];
-      const { promise, cancel } = loader.load(op);
-      let isDone = false;
-      const prevOnce: typeof prev = (result) => {
-        if (isDone) {
-          return;
-        }
-        isDone = true;
-        prev(result);
-      };
-      onDestroy(() => {
-        prevOnce(TRPCClientError.from(new TRPCAbortError(), { isDone: true }));
-        cancel();
+    return ({ op }) => {
+      return observable((observer) => {
+        const loader = loaders[op.type];
+        const { promise, cancel } = loader.load(op);
+
+        promise
+          .then((res) =>
+            observer.next(
+              transformOperationResult(
+                {
+                  meta: res.meta,
+                  data: res.json as any,
+                },
+                runtime.transformer,
+              ),
+            ),
+          )
+          .catch((err) => observer.error(err as any))
+          .finally(() => observer.complete());
+
+        return () => {
+          cancel();
+        };
       });
-      promise
-        .then((envelope) => {
-          prevOnce(transformRPCResponse({ envelope, runtime }));
-        })
-        .catch((cause) => {
-          prevOnce(TRPCClientError.from<TRouter>(cause));
-        });
     };
   };
 }
