+++ packages/client/src/links/httpUtils.ts
@@ -1,6 +1,6 @@
 import { ProcedureType } from '@trpc/server';
-import { PromiseAndCancel } from '../links/core';
-import { LinkRuntimeOptions } from './core';
+import { PromiseAndCancel } from './core';
+import { LinkRuntimeOptions } from './core2';
 
 export interface HTTPLinkOptions {
   url: string;
