+++ packages/client/src/rx/observable.test.ts
@@ -3,9 +3,9 @@ import { waitFor } from '@testing-library/dom';
 import { AnyRouter, TRPCError } from '../../../server/src';
 import { EventEmitter } from 'events';
 import { expectTypeOf } from 'expect-type';
-import { createChain, OperationLink, TRPCLink } from '../links2/core';
-import { dedupeLink } from '../links2/dedupeLink';
-import { splitLink } from '../links2/splitLink';
+import { createChain, OperationLink, TRPCLink } from '../links/core2';
+import { dedupeLink } from '../links/dedupeLink';
+import { splitLink } from '../links/splitLink';
 import { observable } from './observable';
 import { error, map } from './operators';
 import { share } from './operators/share';
