+++ packages/server/test/_testHelpers.ts
@@ -6,7 +6,7 @@ import {
   createWSClient,
   WebSocketClientOptions,
   TRPCWebSocketClient,
-} from '../../client/src/links2/wsLink';
+} from '../../client/src/links/wsLink';
 import ws from 'ws';
 import { createTRPCClient, CreateTRPCClientOptions } from '../../client/src';
 import { AnyRouter, CreateHttpHandlerOptions } from '../src';
