+++ packages/server/test/links.test.ts
@@ -7,11 +7,11 @@ import {
   createChain,
   LinkRuntimeOptions,
   OperationLink,
-} from '../../client/src/links2/core';
-import { httpBatchLink } from '../../client/src/links2/httpBatchLink';
-import { httpLink } from '../../client/src/links2/httpLink';
-import { loggerLink } from '../../client/src/links2/loggerLink';
-import { retryLink } from '../../client/src/links2/retryLink';
+} from '@trpc/client/src/links/core2';
+import { httpBatchLink } from '../../client/src/links/httpBatchLink';
+import { httpLink } from '../../client/src/links/httpLink';
+import { loggerLink } from '../../client/src/links/loggerLink';
+import { retryLink } from '../../client/src/links/retryLink';
 import { observable, toPromise } from '../../client/src/rx/observable';
 import * as trpc from '../src';
 import { AnyRouter } from '../src';
