+++ packages/server/test/transformer.test.ts
@@ -4,8 +4,8 @@ import fetch from 'node-fetch';
 import superjson from 'superjson';
 import { z } from 'zod';
 import { TRPCClientError } from '../../client/src';
-import { httpBatchLink } from '../../client/src/links2/httpBatchLink';
-import { httpLink } from '../../client/src/links2/httpLink';
+import { httpBatchLink } from '../../client/src/links/httpBatchLink';
+import { httpLink } from '../../client/src/links/httpLink';
 import * as trpc from '../src';
 import { TRPCError } from '../src/TRPCError';
 import { routerToServerAndClient, waitError } from './_testHelpers';
